% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateFunctions.R
\name{chol.update}
\alias{chol.update}
\title{EnKF Update with Cholesky Decomposition of Covariance Matrix}
\usage{
\method{chol}{update}(x.mult.prior, y.mult.i, chol.mat, H.mat, tau.mat)
}
\arguments{
\item{x.mult.prior}{Ensemble of state data vectors at time t-1 to update}

\item{y.mult.i}{Ensemble of observation data at time t}

\item{chol.mat}{Cholesky decomposition of the state ensemble covariance matrix}

\item{H.mat}{The observation matrix relating $x_t$ to $y_t$}

\item{tau.mat}{The observation noise matrix}
}
\value{
This function returns the updated state ensemble matrix
}
\description{
This function is given an ensemble state matrix, an ensemble observation matrix,
and the Cholesky decomposition of the estimated state covariance matrix, and 
it updates the ensemble matrix.
}
\examples{
data(x.mult.prior)
data(y.mult.i)
data(H.mat)
data(tau.mat)
sig.mat = chol(cov(t(x.mult.prior)))

chol.update(x.mult.prior, y.mult.i, sig.mat, H.mat, tau.mat)

}
