% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DeltaLorenz04M2Cpp}
\alias{DeltaLorenz04M2Cpp}
\title{A single iteration of the Lorenz 04 model}
\usage{
DeltaLorenz04M2Cpp(X0, F, K, dt, M, newAlgo)
}
\arguments{
\item{dt}{time step}

\item{M}{the number of steps that make a time step}

\item{XX}{initial state}

\item{F_Lor}{F from the Lorenz 04 model}

\item{K_Lor}{K from the Lorenz 04 model}

\item{order}{the order of RK method to use}
}
\value{
dx the increment to reach the new state
}
\description{
A single iteration of the Lorenz 04 model
}
