% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exactGrad}
\alias{exactGrad}
\title{Calculate the exact gradient in the Lorenz model}
\usage{
exactGrad(X, K)
}
\arguments{
\item{X}{the vector at which to evaluate the gradient}

\item{K}{parameter from the Lorenz model}
}
\value{
grad the matrix with the gradient
}
\description{
Calculate the exact gradient in the Lorenz model
}
