% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz.R
\name{Lorenz04M2}
\alias{Lorenz04M2}
\title{Single Lorenz simulation}
\usage{
Lorenz04M2(X0, F_Lor, K_Lor, dt = 0.005, M = 40, order = 1)
}
\arguments{
\item{X0}{The vector of points you want to pass through a Lorenz chaos model.}

\item{F_Lor}{A Forcing constant for the chaos model. Recommended values are}

\item{K_Lor}{A constant that determines the wavelength within the model. A recommended 
value is the length of your data vector divided by 30.}

\item{dt}{The internal step size. The default value is set to 0.005}

\item{M}{The number of internal steps. The default value is set to 40}

\item{order}{The order of the RK solution method to use.
The default value is 1}
}
\value{
The function returns a new vector that represents a single step in a Lorenz chaos
model as described in the paper above.
}
\description{
This function simulates one iteration through a Lorenz chaos model 
described as the second model in 
\href{https://journals.ametsoc.org/doi/pdf/10.1175/JAS3430.1}{Lorenz (2005)}
}
\examples{
X0 = rnorm(960)
Lorenz04M2(X0, 10, 32, 0.005, 40)

}
