% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz.R
\name{Lorenz04M2Sim}
\alias{Lorenz04M2Sim}
\title{Lorenz Model Simulation}
\usage{
Lorenz04M2Sim(X0, F_Lor, K_Lor, dt, M, iter = 500, burn = 100,
  order = 1)
}
\arguments{
\item{X0}{The vector of points you want to pass through a Lorenz chaos model.}

\item{F_Lor}{A Forcing constant for the chaos model. Recommended values are}

\item{K_Lor}{A constant that determines the wavelength within the model. A recommended 
value is the length of your data vector divided by 30.}

\item{dt}{The internal step size. The default value is set to 0.005}

\item{M}{The number of internal steps. The default value is set to 40}

\item{iter}{The number of iterations of the Lorenz model you want to run. 
The default value is set to 500}

\item{burn}{The amount of burn-in you want to run based on your starting point.
The default value is set to 100}

\item{order}{The order of the RK solution method to use.
The default value is 1}
}
\value{
This function returns a matrix containg iter serial simulations of 
Lorenz chaos model updates.
}
\description{
This function simulates a specified number of iterations through a Lorenz chaos model 
described as the second model in 
\href{https://journals.ametsoc.org/doi/pdf/10.1175/JAS3430.1}{Lorenz (2005)}
}
\examples{
X0 = rnorm(960)
Lorenz04M2Sim(X0, 10, 32, 0.005, 40)

}
