% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Lorenz04M2SimCpp}
\alias{Lorenz04M2SimCpp}
\title{Simulate from the Lorenz 04 model}
\usage{
Lorenz04M2SimCpp(Xinit, F_Lor, K_Lor, dt, M, iter, burn, newAlgo)
}
\arguments{
\item{Xinit}{initial state}

\item{F_Lor}{F from the Lorenz 04 model}

\item{K_Lor}{K from the Lorenz 04 model}

\item{dt}{time step}

\item{M}{the number of steps that make a time step}

\item{iter}{number of iterations}

\item{burn}{how many steps to discard}
}
\value{
Xiter values of all iterations
}
\description{
Simulate from the Lorenz 04 model
}
