% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lorenz480}
\alias{Lorenz480}
\title{Lorenz Model N = 480}
\format{A matrix with 500 columns for the simulations and 480 rows for the locations}
\source{
\href{https://journals.ametsoc.org/doi/pdf/10.1175/JAS3430.1}{Lorenz (2005)}
}
\usage{
Lorenz480
}
\description{
500 simulations after a burn-in of 200 from a Lorenz04M2 Model with N = 480, F_Lor = 10, K_Lor = 16,
dt = 0.005, and M = 40. Figure 4(b) gives example of what this model is supposed to
look similar to, up to 960 locations.
}
\keyword{datasets}
