% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RHS}
\alias{RHS}
\title{right hand side of the Lorenz model in vector form}
\usage{
RHS(X, K, F)
}
\arguments{
\item{X}{current state}

\item{K}{parameter from the Lorenz 04 model}

\item{F}{parameter from the Lorenz 04 model}
}
\description{
right hand side of the Lorenz model in vector form
}
