% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S2d.ord}
\alias{S2d.ord}
\title{Example Spatial Location Matrix}
\format{25x2 matrix}
\usage{
S2d.ord
}
\description{
Spatial locations for a 5x5 state-space model for examples ordered using the 
max-min technique.
}
\keyword{datasets}
