% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateFunctions.R
\name{simple.update}
\alias{simple.update}
\title{EnKF Update with Covariance Matrix}
\usage{
simple.update(x.mult.prior, y.mult.i, sigma.mat, H.mat, tau.mat)
}
\arguments{
\item{x.mult.prior}{Ensemble of state data vectors at time t-1 to update}

\item{y.mult.i}{Ensemble of observation data at time t}

\item{sigma.mat}{The covariance matrix for the state ensemble}

\item{H.mat}{The observation matrix relating $x_t$ to $y_t$}

\item{tau.mat}{The observation noise matrix}
}
\value{
This function returns an updated state ensemble matrix
}
\description{
This function is given an ensemble state matrix, an ensemble observation matrix,
and the estimated state covariance matrix, and it updates the ensemble matrix.
}
\examples{
data(x.mult.prior)
data(y.mult.i)
data(H.mat)
data(tau.mat)
sig.mat = cov(t(x.mult.prior))

simple.update(x.mult.prior, y.mult.i, sigma.mat, H.mat, tau.mat)

}
