% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateFunctions.R
\name{tap.update}
\alias{tap.update}
\title{EnKF Update with Tapered Covariance Matrix}
\usage{
tap.update(x.mult.prior, y.mult.i, sigma.mat, H.mat, tau.mat, taper.radius)
}
\arguments{
\item{x.mult.prior}{Ensemble of state data vectors at time t-1 to update}

\item{y.mult.i}{Ensemble of observation data at time t}

\item{sigma.mat}{The estimated state ensemble covariance matrix}

\item{H.mat}{The observation matrix relating $x_t$ to $y_t$}

\item{tau.mat}{The observation noise matrix}

\item{taper.radius}{The radius used to determine the tapered covariance matrix}
}
\value{
This function returns the updated state ensemble matrix
}
\description{
This function is given an ensemble state matrix, an ensemble observation matrix,
and the estimated state covariance matrix, and it tapers the covariance matrix 
before updating the ensemble matrix.
}
\examples{

data(x.mult.prior)
data(y.mult.i)
data(H.mat)
data(tau.mat)
sig.mat = cov(t(x.mult.prior))

tap.update(x.mult.prior, y.mult.i, sig.mat, H.mat, tau.mat, taper.radius = 0.3)

}
