% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tau.mat}
\alias{tau.mat}
\title{Example Observation Noise Matrix}
\format{25x25 matrix}
\usage{
tau.mat
}
\description{
Observation noise matrix for a 5x5 state-space model for examples.
}
\keyword{datasets}
