% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VecchiaUpdate.R
\name{vec.update}
\alias{vec.update}
\title{EnKF update function with Vecchia covariance approximation}
\usage{
vec.update(x.mult.prior, y.mult.i, H.mat, tau.mat, S, eps = 0.001,
  NN = 30, m = NULL)
}
\arguments{
\item{x.mult.prior}{Ensemble of state data vectors at time t-1 to update}

\item{y.mult.i}{Ensemble of observation data at time t}

\item{H.mat}{The observation matrix relating $x_t$ to $y_t$}

\item{tau.mat}{The observation noise matrix}

\item{S}{The matrix of spatial locations in Max-min ordering}

\item{eps}{The threshold for selecting m based upon our hyperpriors. The default value 
is 0.001, and we do not recommend a value larger than 0.5 or smaller than 0.000001.}

\item{NN}{The maximum size wanted for creating a Nearest Neighbors array. This will 
put an upper limit on the number of possible neighbors the method will consider. 
The default value is 30.}

\item{m}{The number of neighbors to consider in the regression models for calculating
the modified Cholesky decomposition of the Covariance matrix. If left \code{NULL},
the method will select m as described above under the threshold \code{eps}. The default
value is set to \code{NULL}.}
}
\value{
A list of returned objects
\itemize{
\item{update}{The updated ensemble of state vectors}
\item{m}{The number of nearest neighbors used to determine the update}
}
}
\description{
Given an observation vector and an ensemble of state vectors, this function returns 
an updated ensemble of state vectors and the number of nearest neighbors used to
determine the update.
}
\examples{
data(x.mult.prior)
data(y.mult.i)
data(H.mat)
data(tau.mat)
data(S2d.ord)

vec.update(x.mult.prior, y.mult.i, H.mat, tau.mat, S2d.ord)
# Choose the m you want
vec.update(x.mult.prior, y.mult.i, H.mat, tau.mat, S2d.ord, m = 20)

}
