\name{VEnKF-package}
\alias{VEnKF-package}
\alias{VEnKF}
\docType{package}
\title{
\packageTitle{VEnKF}
}
\description{
\packageDescription{VEnKF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{VEnKF}
\packageIndices{VEnKF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{VEnKF}

Maintainer: \packageMaintainer{VEnKF}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
