% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEnKF.R
\docType{package}
\name{VEnKF}
\alias{VEnKF}
\alias{VEnKF-package}
\title{VEnKF: Ensemble Kalman Filter using the Vecchia approximation}
\description{
This package can be used for filerting large spatio-temporal datasets
}
