% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x.mult.prior}
\alias{x.mult.prior}
\title{Example State Ensemble Matrix}
\format{25x50 matrix}
\usage{
x.mult.prior
}
\description{
Ensemble state matrix for a 5x5 state-space model with ensemble size N=50 for
examples.
}
\keyword{datasets}
