% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y.mult.i}
\alias{y.mult.i}
\title{Example Observation Ensemble Matrix}
\format{25x50 matrix}
\usage{
y.mult.i
}
\description{
Ensemble observation matrix for a 5x5 state-space model with ensemble size N=50
for examples.
}
\keyword{datasets}
